// demo-mcp-21-code-generator-node/input/reference-repos/MSServicehomepass/src/common/utils/enums/task.enum.ts
export enum Etask {
  APM='CAPTURE_APM_QUERY',
  ADDRESS_COMPLEMENT = 'ADDRESS_COMPLEMENT',
  ADDRESS_COMPLEMENT_ERROR = 'ERROR EN CONSUMO ADDRESS_COMPLEMENT',
  CONSULT_ADDRESS = 'CONSULT_ADDRESS',
  CONSULT_STRUCTURES = 'CONSULT_STRUCTURES',
  CONSULT_DETAIL_ADDRESS = 'CONSULT_DETAIL_ADDRESS',
  CONSULT_HOMEPASS = 'CONSULT_HOMEPASS',
  CONSULT_NEIGHBORHOOD_BY_DANE = 'CONSULT_NEIGHBORHOOD_BY_DANE',
  CONSULT_PUT_CLIENT_HOMEPASS = 'CONSULT_PUT_CLIENT_HOMEPASS',
  CONSULT_TRACEABILITY = 'CONSULT_TRACEABILITY',
  CONSULTANDO_ADDRESS_CREARSOLICITUDINSPIRA = 'CONSULTANDO_ADDRESS_CREARSOLICITUDINSPIRA',
  CONSUMED_SERVICE = 'CONSUMED_SERVICE',
  CONSUMO_REQ_CONDIREXACTAB = 'CONSUMO_REQ_CONSULTADIRECCIONEXACTATABULADA',
  CONSUMO_REQ_CONSULTARDIRECCION = 'CONSUMO_REQ_CONSULTADIRECCION',
  CONSUMO_REQ_CONSULTARDIRECCIONGENERAL = 'CONSUMO_REQ_CONSULTADIRECCIONGENERAL',
  CONSUMO_RES_CONDIREXACTAB = 'CONSUMO_RES_CONSULTADIRECCIONEXACTATABULADA',
  CONSUMO_RES_CONSULTARDIRECCION = 'CONSUMO_RES_CONSULTADIRECCION',
  CONSUMO_RES_CONSULTARDIRECCIONGENERAL = 'CONSUMO_RES_CONSULTADIRECCIONGENERAL',
  CREAR_HOMEPASSUC_CD = 'CREAR__BD_HOMEPASSUC_CONSULTADIRECCION',
  CREAR_HOMEPASSUC_CDG = 'CREAR__BD_HOMEPASSUC_CONSULTADIRECCIONGENERAL',
  CREATE = 'CREANDO_MOCKUP',
  CREATE_CUSREQUEST_HOMEPASS_GENERAL = 'CREATE_CUSREQUEST_HOMEPASS_GENERAL',
  CREATE_REQUEST = 'CREATE_REQUEST',
  CREATE_SERVICE_TRACING = 'CREATE_SERVICE_TRACING',
  ERROR_CONS_CONSULT_SP_GENERAL = 'ERROR_CONS_CONSULT_SP_GENERAL',
  ERROR_CONS_PRC_CONSULTA_DIRECCION_SP = 'ERROR_CONS_PRC_CONSULTA_DIRECCION_SP',
  ERROR_GET_PARAM = 'Error obteniendo parametros',
  ERROR_LOAD_PARAM = 'Error cargando parametros',
  ERROR_METODO_HOMEPASSUC = 'ERROR_METODO_HOMEPASSUC',
  ERROR_UPDATE_PARAM = 'Error actualizando parametros',
  EXCEPTION_MANAGER = 'EXCEPTION_MANAGER',
  EXECUTE_SP_ORACLE = 'EXECUTE_SP_ORACLE',
  FIN_HOMEPASS = 'FIN_RESPONSEHOMEPASS',
  FIND_CUSREQUESTHOMEPASS = 'CONSULTANDO CUSREQUESTHOMEPASS',
  FINDALL = 'CONSULTANDO_MOCKUPS',
  FINDONE = 'CONSULTANDO_MOCKUP',
  GET_CUSREQUESTHOMEPASS = 'GET_CUSREQUESTHOMEPASS',
  GET_ERROR_SERVICE = 'GET_ERROR_SERVICE',
  GET_SERVICE_TRACING = 'GET_SERVICE_TRACING',
  GET_STATE_HOMEPASS = 'GET_STATE_HOMEPASS',
  INCIO_HOMEPASS = 'INICIO_REQUESTHOMEPASS',
  INCIO_HOMEPASSUC = 'INICIO_REQUESTHOMEPASSUC',
  JOB_EXECUTION = 'JOB_EXECUTION',
  LOAD_MESSAGE = 'CARGANDO_MENSAJES',
  LOAD_PARAM = 'LOAD_PARAM',
  MANUAL_JOB_EXECUTION = 'MANUAL_JOB_EXECUTION',
  NOT_FOUND_CUSREQUESTHOMEPASS = 'NO SE ENCONTRO REGISTRO EN CUSREQUESTHOMEPASS',
  ORCHESTRATOR = 'ORCHESTRATOR',
  REMOVE = 'ELIMINANDO_MOCKUP',
  REQUEST_CMATRICES_BUSCAR_SOLICITUD_POR_ID_SOLICITUD = 'REQUEST_CMATRICES_BUSCAR_SOLICITUD_POR_ID_SOLICITUD',
  REQUEST_DETAIL_ADDRESS = 'REQUEST_DETAIL_ADDRESS',
  REQUEST_HTTP = 'REQUEST_HTTP',
  REQUEST_HTTP_REST = 'CONSULTANDO API REST',
  REQUEST_MS_APPOINT_ADMIN_CAPACITY = 'REQUEST_MS_APPOINT_ADMIN_CAPACITY',
  REQUEST_MS_APPOINT_ADMIN_CREATE_ORDER = 'REQUEST_MS_APPOINT_ADMIN_CREATE_ORDER',
  REQUEST_MS_RES_ORDE_RESERVE_SEND_ORDER = 'MSRES_ORDE_RESERVE_SEND_ORDER',
  RESPONSE_CONS_PRC_CONSULTA_DIRECCION_SP = 'RESPONSE_CONS_PRC_CONSULTA_DIRECCION_SP',
  RESPONSE_DETAIL_ADDRESS = 'RESPONSE_DETAIL_ADDRESS',
  SERVICE_HOMEPASS = 'SERVICE_HOMEPASS',
  UPDATE_WITH_IDCASETCRM = 'ACTUALIZANDO_BASE_DE_DATOS_',
  VALIDATE_CHANNEL = 'VALIDATE_CHANNEL',
  SEGMENTATION_BY_TYPE = 'SEGMENTATION_BY_TYPE',
  GET_STATEHOMEPASS = 'GET_STATEHOMEPASS',
  XML_TO_JSON = 'XML_TO_JSON',
  ORACLE_DATABASE = 'ORACLE_DATABASE'
}